+++ src/mixins/helpers.js
@@ -4,16 +4,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {getTrackCSS, getTrackLeft, getTrackAnimateCSS} from './trackHelper';
 import assign from 'object-assign';
-import { getOnDemandLazySlides } from '../utils/innerSliderUtils'
+import { getOnDemandLazySlides, getWidth, getHeight, canGoNext } from '../utils/innerSliderUtils'
 
 var helpers = {
-  // supposed to start autoplay of slides
-  initialize: function (props) {
+  update: function (props, recursionLevel=0) {
     const slickList = ReactDOM.findDOMNode(this.list);
-
     var slideCount = React.Children.count(props.children);
-    var listWidth = this.getWidth(slickList);
-    var trackWidth = this.getWidth(ReactDOM.findDOMNode(this.track));
+    var listWidth = getWidth(slickList);
+    var trackWidth = getWidth(ReactDOM.findDOMNode(this.track));
     var slideWidth;
 
     if (!props.vertical) {
@@ -21,58 +19,12 @@ var helpers = {
       if (props.centerPadding.slice(-1) === '%') {
         centerPaddingAdj *= listWidth / 100
       }
-      slideWidth = Math.ceil((this.getWidth(ReactDOM.findDOMNode(this)) - centerPaddingAdj)/props.slidesToShow)
+      slideWidth = Math.ceil((getWidth(slickList) - centerPaddingAdj)/props.slidesToShow)
     } else {
-      slideWidth = Math.ceil(this.getWidth(ReactDOM.findDOMNode(this)))
+      slideWidth = Math.ceil(getWidth(slickList))
     }
 
-    const slideHeight = this.getHeight(slickList.querySelector('[data-index="0"]'));
-    const listHeight = slideHeight * props.slidesToShow;
-
-    var currentSlide = props.rtl ? slideCount - 1 - props.initialSlide : props.initialSlide;
-
-    this.setState({
-      slideCount,
-      slideWidth,
-      listWidth,
-      trackWidth,
-      currentSlide,
-      slideHeight,
-      listHeight,
-    }, function () {
-      // this reference isn't lost due to mixin
-      var targetLeft = getTrackLeft(assign({
-        slideIndex: this.state.currentSlide,
-        trackRef: this.track
-      }, props, this.state));
-      // getCSS function needs previously set state
-      var trackStyle = getTrackCSS(assign({left: targetLeft}, props, this.state));
-
-      this.setState({trackStyle: trackStyle});
-
-      this.autoPlay(); // once we're set up, trigger the initial autoplay.
-    });
-  },
-  update: function (props) {
-    const slickList = ReactDOM.findDOMNode(this.list);
-    // This method has mostly same code as initialize method.
-    // Refactor it
-    var slideCount = React.Children.count(props.children);
-    var listWidth = this.getWidth(slickList);
-    var trackWidth = this.getWidth(ReactDOM.findDOMNode(this.track));
-    var slideWidth;
-
-    if (!props.vertical) {
-      var centerPaddingAdj = props.centerMode && (parseInt(props.centerPadding) * 2);
-      if (props.centerPadding.slice(-1) === '%') {
-        centerPaddingAdj *= listWidth / 100
-      }
-      slideWidth = Math.ceil((this.getWidth(ReactDOM.findDOMNode(this)) - centerPaddingAdj)/props.slidesToShow)
-    } else {
-      slideWidth = Math.ceil(this.getWidth(ReactDOM.findDOMNode(this)))
-    }
-
-    const slideHeight = this.getHeight(slickList.querySelector('[data-index="0"]'));
+    const slideHeight = getHeight(slickList.querySelector('[data-index="0"]'));
     const listHeight = slideHeight * props.slidesToShow;
 
     // pause slider if autoplay is set to false
@@ -96,7 +48,11 @@ var helpers = {
       listHeight,
       lazyLoadedList: prevLazyLoadedList.concat(slidesToLoad)
     }, function () {
-
+      if (!slideWidth) {
+        if (recursionLevel < 2) {
+          this.update(this.props, recursionLevel + 1)
+        }
+      }
       var targetLeft = getTrackLeft(assign({
         slideIndex: this.state.currentSlide,
         trackRef: this.track
@@ -107,12 +63,6 @@ var helpers = {
       this.setState({trackStyle: trackStyle});
     });
   },
-  getWidth: function getWidth(elem) {
-    return elem && (elem.getBoundingClientRect().width || elem.offsetWidth) || 0;
-  },
-  getHeight(elem) {
-    return elem && (elem.getBoundingClientRect().height || elem.offsetHeight) || 0;
-  },
   adaptHeight: function () {
     if (this.props.adaptiveHeight) {
       var selector = '[data-index="' + this.state.currentSlide +'"]';
@@ -122,24 +72,6 @@ var helpers = {
         slickList.style.height = (elem.offsetHeight || 0) + 'px';
       }
     }
-  },
-  canGoNext: function (opts){
-    var canGo = true;
-    if (!opts.infinite) {
-      if (opts.centerMode) {
-        // check if current slide is last slide
-        if (opts.currentSlide >= (opts.slideCount - 1)) {
-          canGo = false;
-        }
-      } else {
-        // check if all slides are shown in slider
-        if (opts.slideCount <= opts.slidesToShow ||
-          opts.currentSlide >= (opts.slideCount - opts.slidesToShow)) {
-          canGo = false;
-        }
-      }
-    }
-    return canGo;
   },
   slideHandler: function (index) {
     // index is target slide index
@@ -190,7 +122,7 @@ var helpers = {
       this.setState({
         animating: true,
         currentSlide: animationTargetSlide
-      }, function () {
+      }, () => {
         if (this.props.asNavFor && this.props.asNavFor.innerSlider.state.currentSlide !== this.state.currentSlide) {
           this.props.asNavFor.innerSlider.slideHandler(index)
         }
@@ -314,12 +246,11 @@ var helpers = {
           delete this.animationEndCallback;
         });
       };
-
       this.setState({
         animating: true,
         currentSlide: finalTargetSlide,
         trackStyle: getTrackAnimateCSS(assign({left: animationTargetLeft}, this.props, this.state))
-      }, function () {
+      }, () => {
         if (this.props.asNavFor && this.props.asNavFor.innerSlider.state.currentSlide !== this.state.currentSlide) {
           this.props.asNavFor.innerSlider.slideHandler(index)
         }
@@ -329,33 +260,6 @@ var helpers = {
     }
 
     this.autoPlay();
-  },
-  swipeDirection: function (touchObject) {
-    var xDist, yDist, r, swipeAngle;
-
-    xDist = touchObject.startX - touchObject.curX;
-    yDist = touchObject.startY - touchObject.curY;
-    r = Math.atan2(yDist, xDist);
-
-    swipeAngle = Math.round(r * 180 / Math.PI);
-    if (swipeAngle < 0) {
-        swipeAngle = 360 - Math.abs(swipeAngle);
-    }
-    if ((swipeAngle <= 45) && (swipeAngle >= 0) || (swipeAngle <= 360) && (swipeAngle >= 315)) {
-        return 'left';
-    }
-    if ((swipeAngle >= 135) && (swipeAngle <= 225)) {
-        return 'right';
-    }
-    if (this.props.verticalSwiping === true) {
-      if ((swipeAngle >= 35) && (swipeAngle <= 135)) {
-        return 'down';
-      } else {
-        return 'up';
-      }
-    }
-
-    return 'vertical';
   },
   play: function(){
     var nextIndex;
@@ -367,7 +271,7 @@ var helpers = {
     if (this.props.rtl) {
       nextIndex = this.state.currentSlide - this.props.slidesToScroll;
     } else {
-      if (this.canGoNext(Object.assign({}, this.props,this.state))) {
+      if (canGoNext(Object.assign({}, this.props,this.state))) {
         nextIndex = this.state.currentSlide + this.props.slidesToScroll;
       } else {
         return false;